[CHERRY-PICK] Adding 'bedc48728e20b66dee054f1a68895cca8e8913a5' to 'CHERRY_PICK_main_2024-02-22_16.26.31' using the python script version 'v1.0.6'.

+++ PresentationLayer/ACCOUNT/AccountingVouchers.aspx.cs
@@ -17233,7 +17233,7 @@ public partial class AccountingVouchers : System.Web.UI.Page
         dtotheractivity.Columns.Add(new DataColumn("CostCenterID", typeof(int)));
         dtotheractivity.Columns.Add(new DataColumn("PartyID", typeof(int)));
         dtotheractivity.Columns.Add(new DataColumn("PartyName", typeof(string)));
-        dtotheractivity.Columns.Add(new DataColumn("Amount", typeof(int)));
+        dtotheractivity.Columns.Add(new DataColumn("Amount", typeof(decimal)));
 
         return dtotheractivity;
     }
